Fixed Modal content render bug, created Modal content Components, created onChange functionality
